Add files via upload
This notebook is about Task 03 of our data challenge, which includes the approaches along with the corrected images. Also, you can find that we have used slider to check all the images at once. Finally, at the end, we have also added the code for merging into a video along with false colour.